chore: enable chart zoom
